Move the warnset commands into another file
